+++ memrchr_inv.c
@@ -1,5 +1,5 @@
 /* See LICENSE file for copyright and license details. */
-#include "libsimple.h"
+#include "common.h"
 #ifndef TEST
 
 
@@ -7,7 +7,7 @@ void *
 libsimple_memrchr_inv(const void *s_, int c_, size_t n_)
 {
 	char *s = *(char **)(void *)&s_, c = (char)c_;
-	ssize_t n = n_;
+	ssize_t n = (ssize_t)n_;
 	while (n-- && s[n] == c);
 	return n < 0 ? NULL : &s[n];
 }
